feat: move employee assignments to hr views
