+++ locales/it/administration.json
@@ -59,6 +59,7 @@
       "internal": "Interni",
       "external": "Esterni",
       "costs": "Costi",
+      "employee_assignments": "Assegnazioni Dipendenti",
       "work_units": "Unità di Lavoro",
       "work_units_all": "Unità di Lavoro (Tutte)"
     },
