+++ utils/permissions.ts
@@ -62,6 +62,7 @@ export const PERMISSION_DEFINITIONS: PermissionDefinition[] = [
   { id: 'hr.internal', actions: CRUD, module: 'hr' },
   { id: 'hr.external', actions: CRUD, module: 'hr' },
   { id: 'hr.costs', actions: VIEW_UPDATE, module: 'hr' },
+  { id: 'hr.employee_assignments', actions: ['update'], module: 'hr' },
   { id: 'hr.work_units', actions: CRUD, module: 'hr' },
   { id: 'hr.work_units_all', actions: VIEW_ONLY, isScope: true, module: 'hr' },
